refactor: moved routine to a new file
